% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataInThePackage.R
\docType{data}
\name{Age (from ExampleData_Liver.rda)}
\alias{Age (from ExampleData_Liver.rda)}
\alias{Age}
\title{ExampleData_Liver.rda}
\format{
A vecter including the chronological age.
}
\usage{
data(ExampleData_Liver)
}
\description{
The sample liver data from GSE107038 contains 39 normal liver samples. Only
CTS clock CpGs are in the data to control data size.
}
\keyword{datasets}
