% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataInThePackage.R
\docType{data}
\name{beta.m (from MurphyGSE88890.rda)}
\alias{beta.m (from MurphyGSE88890.rda)}
\alias{beta.m}
\alias{phenotype.df}
\title{MurphyGSE88890.rda}
\format{
A DNAm matrix. Row: CpGs, column: samples.

A data frame including the chronological age.
}
\usage{
data(MurphyGSE88890)

data(MurphyGSE88890)
}
\description{
The example bulk prefrontal cortex (PFC) data is from Murphy et al. This is an
Illumina 450k dataset encompassing 38 normal PFC tissue samples and 37 PFC
major depressive disorder samples. Only controls are used here and only CTS
clock CpGs are in the data to control data size.

The example bulk prefrontal cortex (PFC) data is from Murphy et al. This is an
Illumina 450k dataset encompassing 38 normal PFC tissue samples and 37 PFC
major depressive disorder samples. Only controls are used here and only CTS
clock CpGs and the necesarry deconvolution CpGs are in the data to control data size.
}
\keyword{datasets}
