% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataInThePackage.R
\docType{data}
\name{beta.m (from PaiGSE112179.rda)}
\alias{beta.m (from PaiGSE112179.rda)}
\alias{beta.m}
\alias{phenotype.df}
\title{PaiGSE112179.rda}
\format{
A DNAm matrix. Row: CpGs, column: samples.

A data frame including the chronological age.
}
\usage{
data(PaiGSE112179)

data(PaiGSE112179)
}
\description{
The sorted PFC neuron data is from Pai et al. This is an Illumina EPIC dataset
encompassing 33 normal neuron tissue samples, 35 schizophrenia samples and
32 bipolar disorder neuron samples. Only controls are used here and only CTS
clock CpGs and the necesarry deconvolution CpGs are in the data to control data size.

The sorted PFC neuron data is from Pai et al. This is an Illumina EPIC dataset
encompassing 33 normal neuron tissue samples, 35 schizophrenia samples and
32 bipolar disorder neuron samples. Only controls are used here and only CTS
clock CpGs are in the data to control data size.
}
\keyword{datasets}
