% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataInThePackage.R
\docType{data}
\name{clockCoef.df (from BrainCoef.rda)}
\alias{clockCoef.df (from BrainCoef.rda)}
\alias{clockCoef.df}
\title{BrainCoef.rda}
\format{
A data frame including 'coef', coefficients, and 'probe', CpGs.
}
\usage{
data(BrainCoef)
}
\description{
The clock's CpGs and coefficients.
}
\keyword{datasets}
