% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataInThePackage.R
\docType{data}
\name{clockCoef.df (from Glia-ExCoef.rda)}
\alias{clockCoef.df (from Glia-ExCoef.rda)}
\alias{clockCoef.df}
\title{Glia-ExCoef.rda}
\format{
A data frame including 'coef', coefficients, and 'probe', CpGs
}
\usage{
data(Glia-ExCoef)
}
\description{
The clock's CpGs and coefficients.
}
\keyword{datasets}
