% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataInThePackage.R
\docType{data}
\name{clockCoef.df (from Neu-InCoef.rda)}
\alias{clockCoef.df (from Neu-InCoef.rda)}
\alias{clockCoef.df}
\title{Neu-InCoef.rda}
\format{
A data frame including 'coef', coefficients, and 'probe', CpGs
}
\usage{
data(Neu-InCoef)
}
\description{
The clock's CpGs and coefficients.
}
\keyword{datasets}
