% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataInThePackage.R
\docType{data}
\name{HepClock.glm (from HepClock.rda)}
\alias{HepClock.glm (from HepClock.rda)}
\alias{HepClock.glm}
\title{HepClock.rda}
\format{
An object with S3 class generated by glmnet::glmnet().
}
\usage{
data(HepClock)
}
\description{
An object with S3 class generated by glmnet which include the
coefficients and CpGs of the clock.
}
\keyword{datasets}
