% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataInThePackage.R
\docType{data}
\name{LiverClock.glm (from LiverClock.rda)}
\alias{LiverClock.glm (from LiverClock.rda)}
\alias{LiverClock.glm}
\title{LiverClock.rda}
\format{
An object with S3 class generated by glmnet::glmnet().
}
\usage{
data(LiverClock)
}
\description{
An object with S3 class generated by glmnet which include the
coefficients and CpGs of the clock.
}
\keyword{datasets}
